+++ CHANGELOG.md
@@ -5,10 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0][] - 2023-02-23
+
+### Added
+
+- Added ability to show an arrival card
+- Added support for landmark routing
+- Added support for voice directions
+- Added ability to show remaining distance to points of interest
+- Added ability to show remaining walk time to points of interest
+
 ## [1.4.0][] - 2022-08-22
 
 ### Added
 
 - Initial release
 
+[1.7.0]: https://github.com/phunware/maas-mapping-module-ios/compare/1.4.0...1.7.0
 [1.4.0]: https://github.com/phunware/maas-mapping-module-ios/tree/1.4.0